+++ release/plugins/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
-require('./main/index.js');
-require('./schema/index.js');
-require('./request/index.js');
-require('./utils/index.js');
 require('./handler/index.js');
+require('./main/index.js');
+require('./request/index.js');
+require('./schema/index.js');
+require('./utils/index.js');