+++ release/plugins/main/index.js
@@ -1,16 +1,16 @@
 'use strict';
 
-require('./server/index.js');
 require('./api/index.js');
+require('./body/index.js');
 require('./contact/index.js');
-require('./license/index.js');
-require('./host/index.js');
-require('./variable/index.js');
-require('./tag/index.js');
 require('./endpoint/index.js');
+require('./host/index.js');
+require('./license/index.js');
+require('./param/index.js');
+require('./preset/index.js');
 require('./response/index.js');
 require('./return/index.js');
-require('./param/index.js');
-require('./body/index.js');
-require('./preset/index.js');
+require('./server/index.js');
+require('./tag/index.js');
+require('./variable/index.js');