+++ release/plugins/request/index.js
@@ -1,11 +1,11 @@
 'use strict';
 
-require('./success/index.js');
+require('./cms/index.js');
+require('./cookie/index.js');
 require('./error/index.js');
+require('./file/index.js');
+require('./header/index.js');
 require('./proxy/index.js');
 require('./redirect/index.js');
-require('./header/index.js');
-require('./cookie/index.js');
-require('./file/index.js');
-require('./cms/index.js');
+require('./success/index.js');