+++ release/plugins/schema/integer/integer.js
@@ -16,7 +16,7 @@ var optional = require('../../../utils/rules/optional/optional.js');
 var pipe = require('../../../utils/rules/pipe/pipe.js');
 
 const integer = () => {
-    const { default: defaultValue, example, examples, format = 'int32', max: max$1, min: min$1, values: values$1, ...props } = jsx.useProps() || {};
+    const { default: defaultValue, example, examples, exclusiveMaximum, exclusiveMinimum, format = 'int32', max: max$1, min: min$1, multipleOf, values: values$1, ...props } = jsx.useProps() || {};
     const schema = useSchemaType.useSchemaType('integer', {
         ...props,
         default: defaultValue !== undefined ? Number(defaultValue) : undefined,
@@ -24,12 +24,26 @@ const integer = () => {
         examples: examples === null || examples === void 0 ? void 0 : examples.map(Number),
         values: values$1 === null || values$1 === void 0 ? void 0 : values$1.map(Number),
     });
-    // @ts-expect-error: FIXME
-    schema.format = format;
-    // @ts-expect-error: FIXME
-    schema.minimum = min$1 !== undefined ? Number(min$1) : undefined;
-    // @ts-expect-error: FIXME
-    schema.maximum = max$1 !== undefined ? Number(max$1) : undefined;
+    if (schema) {
+        if (format) {
+            schema.format = format;
+        }
+        if (min$1 !== undefined) {
+            schema.minimum = Number(min$1);
+        }
+        if (max$1 !== undefined) {
+            schema.maximum = Number(max$1);
+        }
+        if (exclusiveMinimum) {
+            schema.exclusiveMinimum = typeof exclusiveMinimum === 'boolean' ? exclusiveMinimum : Number(exclusiveMinimum);
+        }
+        if (exclusiveMaximum) {
+            schema.exclusiveMaximum = typeof exclusiveMaximum === 'boolean' ? exclusiveMaximum : Number(exclusiveMaximum);
+        }
+        if (multipleOf !== undefined) {
+            schema.multipleOf = Number(multipleOf);
+        }
+    }
     const rules = [];
     if (defaultValue !== undefined) {
         rules.push(defaultTo.defaultTo(defaultValue));