+++ release/plugins/schema/number/number.js
@@ -17,12 +17,28 @@ var useRule = require('../../../hooks/useRule/useRule.js');
 var pipe = require('../../../utils/rules/pipe/pipe.js');
 
 const number = () => {
-    const { max: max$1, min: min$1, ...props } = jsx.useProps() || {};
+    const { exclusiveMaximum, exclusiveMinimum, format, max: max$1, min: min$1, multipleOf, ...props } = jsx.useProps() || {};
     const schema = useSchemaType.useSchemaType('number', props);
-    // @ts-expect-error: FIXME
-    schema.minimum = min$1;
-    // @ts-expect-error: FIXME
-    schema.maximum = max$1;
+    if (schema) {
+        if (format !== undefined) {
+            schema.format = format;
+        }
+        if (min$1 !== undefined) {
+            schema.minimum = min$1;
+        }
+        if (max$1 !== undefined) {
+            schema.maximum = max$1;
+        }
+        if (exclusiveMinimum) {
+            schema.exclusiveMinimum = typeof exclusiveMinimum === 'boolean' ? exclusiveMinimum : Number(exclusiveMinimum);
+        }
+        if (exclusiveMaximum) {
+            schema.exclusiveMaximum = typeof exclusiveMaximum === 'boolean' ? exclusiveMaximum : Number(exclusiveMaximum);
+        }
+        if (multipleOf !== undefined) {
+            schema.multipleOf = Number(multipleOf);
+        }
+    }
     const rules = [];
     if (props.default !== undefined) {
         rules.push(defaultTo.defaultTo(props.default));