+++ release/plugins/utils/index.es6.js
@@ -1,6 +1,6 @@
-import './swagger/index.es6.js';
+import './blacklist/index.es6.js';
 import './dts/index.es6.js';
 import './env/index.es6.js';
 import './protection/index.es6.js';
-import './blacklist/index.es6.js';
+import './swagger/index.es6.js';
 import './whitelist/index.es6.js';