+++ release/plugins/utils/index.js
@@ -1,9 +1,9 @@
 'use strict';
 
-require('./swagger/index.js');
+require('./blacklist/index.js');
 require('./dts/index.js');
 require('./env/index.js');
 require('./protection/index.js');
-require('./blacklist/index.js');
+require('./swagger/index.js');
 require('./whitelist/index.js');