+++ release/utils/index.js
@@ -1,16 +1,16 @@
 'use strict';
 
-require('./parseSearch/index.js');
-require('./stringifySearch/index.js');
-require('./httpOnStart/index.js');
+require('./FileData/index.js');
+require('./JSONString/index.js');
+require('./action/index.js');
+require('./decorators/index.js');
+require('./generateTypes/index.js');
 require('./getEndpoint/index.js');
 require('./getOrAdd/index.js');
-require('./decorators/index.js');
-require('./action/index.js');
+require('./httpOnStart/index.js');
 require('./parseBody/index.js');
 require('./parseFormBody/index.js');
+require('./parseSearch/index.js');
 require('./rules/index.js');
-require('./JSONString/index.js');
-require('./FileData/index.js');
-require('./generateTypes/index.js');
+require('./stringifySearch/index.js');