+++ release/utils/rules/index.js
@@ -2,37 +2,37 @@
 
 Object.defineProperty(exports, '__esModule', { value: true });
 
-var helpers = require('./helpers.js');
-var constants = require('./constants.js');
-require('./types.js');
-require('./num/index.js');
 require('./arrayOf/index.js');
-require('./oneOf/index.js');
+require('./bin/index.js');
+require('./binaryAccept/index.js');
+var constants = require('./constants.js');
 require('./dateTo/index.js');
+require('./defaultTo/index.js');
+var helpers = require('./helpers.js');
 require('./int/index.js');
+require('./max/index.js');
+require('./maxBin/index.js');
+require('./maxDate/index.js');
+require('./maxLength/index.js');
+require('./min/index.js');
+require('./minBin/index.js');
+require('./minDate/index.js');
+require('./minLength/index.js');
 require('./nullable/index.js');
+require('./num/index.js');
 require('./objectOf/index.js');
+require('./oneOf/index.js');
 require('./optional/index.js');
 require('./pattern/index.js');
+require('./pipe/index.js');
 require('./required/index.js');
 require('./tupleOf/index.js');
+require('./types.js');
 require('./uuidTo/index.js');
 require('./values/index.js');
-require('./pipe/index.js');
-require('./maxDate/index.js');
-require('./minDate/index.js');
-require('./maxLength/index.js');
-require('./minLength/index.js');
-require('./max/index.js');
-require('./min/index.js');
-require('./defaultTo/index.js');
-require('./bin/index.js');
-require('./minBin/index.js');
-require('./maxBin/index.js');
-require('./binaryAccept/index.js');
 
 
 
+exports.rulesErrors = constants.rulesErrors;
 exports.RulesError = helpers.RulesError;
 exports.addKey = helpers.addKey;
-exports.rulesErrors = constants.rulesErrors;