+++ src/plugins/schema/integer/integer.ts
@@ -15,9 +15,51 @@ import {
 } from '../../../utils'
 
 export interface IntegerProps extends ValuesSchemaProps<bigint | number> {
+  /**
+   * The `exclusiveMaximum` keyword is used to restrict the value to be less than the specified number.
+   * @example For example, the following value is valid:
+   * ```tsx
+   * <integer exclusiveMaximum={10} />
+   * ```
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#numbers
+   * */
+  exclusiveMaximum?: bigint | boolean | number
+
+  /**
+   * The `exclusiveMinimum` keyword is used to restrict the value to be greater than the specified number.
+   * @example For example, the following value is valid:
+   * ```tsx
+   * <integer exclusiveMinimum={10} />
+   * ```
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#numbers
+   * */
+  exclusiveMinimum?: bigint | boolean | number
+
+  /**
+   * An optional format modifier serves as a hint at the contents and format of the string.
+   * @example For example, the following value is valid:
+   * ```tsx
+   * <integer format='int64' />
+   * ```
+   * @see https://swagger.io/docs/specification/data-models/data-types/#numbers
+   * */
   format?: IntegerFormats
+
+  /** Validate the integer number value by maximum. */
   max?: bigint | number
+
+  /** Validate the integer number value by minimum. */
   min?: bigint | number
+
+  /**
+   * The `multipleOf` keyword is used to restrict the value to be a multiple of the specified number.
+   * @example For example, the following value is valid:
+   * ```tsx
+   * <number multipleOf={2} />
+   * ```
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#numbers
+   * */
+  multipleOf?: bigint | number
 }
 
 export const integer: HandlerPlugin = () => {
@@ -25,9 +67,12 @@ export const integer: HandlerPlugin = () => {
     default: defaultValue,
     example,
     examples,
+    exclusiveMaximum,
+    exclusiveMinimum,
     format = 'int32',
     max,
     min,
+    multipleOf,
     values,
     ...props
   } = useProps<IntegerProps>() || {}
@@ -38,12 +83,32 @@ export const integer: HandlerPlugin = () => {
     examples: examples?.map(Number),
     values: values?.map(Number),
   })
-  // @ts-expect-error: FIXME
-  schema.format = format
-  // @ts-expect-error: FIXME
-  schema.minimum = min !== undefined ? Number(min) : undefined
-  // @ts-expect-error: FIXME
-  schema.maximum = max !== undefined ? Number(max) : undefined
+
+  if (schema) {
+    if (format) {
+      schema.format = format
+    }
+
+    if (min !== undefined) {
+      schema.minimum = Number(min)
+    }
+
+    if (max !== undefined) {
+      schema.maximum = Number(max)
+    }
+
+    if (exclusiveMinimum) {
+      schema.exclusiveMinimum = typeof exclusiveMinimum === 'boolean' ? exclusiveMinimum : Number(exclusiveMinimum)
+    }
+
+    if (exclusiveMaximum) {
+      schema.exclusiveMaximum = typeof exclusiveMaximum === 'boolean' ? exclusiveMaximum : Number(exclusiveMaximum)
+    }
+
+    if (multipleOf !== undefined) {
+      schema.multipleOf = Number(multipleOf)
+    }
+  }
 
   const rules: Rule[] = []