+++ src/plugins/schema/string/string.ts
@@ -7,14 +7,47 @@ import { type ValuesSchemaProps } from '../../../types'
 import { defaultTo, maxLength, minLength, pattern as patternTo, pipe, type Rule, values } from '../../../utils'
 
 export interface StringProps extends ValuesSchemaProps <string> {
+  /**
+   * An optional format modifier serves as a hint at the contents and format of the string.
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#strings
+   * */
+  // eslint-disable-next-line perfectionist/sort-intersection-types, @typescript-eslint/ban-types
+  format?: 'binary' | 'byte' | 'date' | 'date-time' | 'email' | 'hostname' | 'ipv4' | 'ipv6' | 'password' | 'uri' | 'uuid' | (string & {})
+
+  /**
+   * String length can be restricted using `min` and `max`.
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#strings
+   * */
   max?: number
+
+  /**
+   * String length can be restricted using `min` and `max`.
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#strings
+   * */
   min?: number
+
+  /**
+   * The pattern keyword lets you define a regular expression template for the string value.
+   * Only the values that match this template will be accepted.
+   * The regular expression syntax used is from JavaScript (more specifically, ECMA 262).
+   * Regular expressions are case-sensitive, that is, [a-z] and [A-Z] are different expressions.
+   * @example For example, the following pattern matches a Social Security Number (SSN) in the 123-45-6789 format:
+   * ```tsx
+   * <string pattern='^\d{3}-\d{2}-\d{4}$' />
+   * ```
+   * @see https://swagger.io/docs/specification/v3_0/data-models/data-types/#strings
+   * */
   pattern?: RegExp | string
+
+  /**
+   * The `patternId` keyword is used to reference a pattern from the OpenAPI specification.
+   * */
   patternId?: string
 }
 
 export const string: HandlerPlugin = () => {
   const {
+    format,
     max,
     min,
     pattern,
@@ -27,6 +60,10 @@ export const string: HandlerPlugin = () => {
   if (schema) {
     const rules: Rule[] = []
 
+    if (format !== undefined) {
+      schema.format = format
+    }
+
     if (props.default !== undefined) {
       rules.push(defaultTo(props.default))
     }
@@ -37,13 +74,17 @@ export const string: HandlerPlugin = () => {
       rules.push(values(props.values))
     }
 
+    if (format !== undefined) {
+      schema.format = format
+    }
+
     if (min !== undefined) {
-      schema.minimum = min
+      schema.minLength = min
       rules.push(minLength(min))
     }
 
     if (max !== undefined) {
-      schema.maximum = max
+      schema.maxLength = max
       rules.push(maxLength(max))
     }